always emit session_id as integer
